pro_error_detect/requirements.txt
streamlit
requests
pillow
torch  
